Added monolithic 'Connect-WordPressAccount'
Decided that having the user do a two stage auth was cludgey, so I made
Get-WordPressAuthCode and Get-WordPressAuthToken into private functions.

Also moved the oAuth window into it's own private function.

Added private secure storage of an oAuth Access Token, so that the user
need not register every time.  The token will be automatically detected
when the cmdlets are subsequently reloaded.